Cache updated guild icon and/or name
When a user updates their server icon and/or name, it does not update in the guild cache, which leads to the server listing always showing the old icon and/or name until they refresh the list. This fixes that by checking if the guild icon or name has been changed and updates the cache.

How to check if the code works:
Open the dashboard.
Inspect elements.
Go into application.
Open local storage.
Find the guilds key and change one of the server's icon or name.
Reload.
Click/open the server settings.
Go back.
Reload.
The icon and/or name should be fixed then (hopefully).